Make the M_MAXLOOP algorithm clearer.

+++ kern/uipc_mbuf.c
@@ -265,7 +265,7 @@ void
 m_resethdr(struct mbuf *m)
 {
 	int len = m->m_pkthdr.len;
-	u_int8_t loop = m->m_pkthdr.ph_loop;
+	u_int16_t loop = m->m_pkthdr.ph_loop;
 
 	KASSERT(m->m_flags & M_PKTHDR);
 	m->m_flags &= (M_EXT|M_PKTHDR|M_EOR|M_EXTWR|M_ZEROIZE);
